# src/backend/api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Trading Bot API
  description: API documentation for the trading bot system
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.tradingbot.com/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /strategies/{strategy_id}:
    get:
      summary: Get strategy details
      parameters:
        - in: path
          name: strategy_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Strategy details
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Strategy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
    put:
      summary: Update strategy
      parameters:
        - in: path
          name: strategy_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Strategy updated
        '400':
          description: Invalid strategy configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Strategy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
    delete:
      summary: Delete strategy
      parameters:
        - in: path
          name: strategy_id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Strategy deleted
        '404':
          description: Strategy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...